[new] 에디터 분리
