+++ src/components/common/Layout/Layout.module.scss
@@ -3,7 +3,7 @@
 
   &-cont {
     width: 1060px;
-    height: calc(100% - 100px);
+    min-height: calc(100% - 100px);
     padding: var(--spacing-20);
     margin: 0 auto;
   }
